+++ CHANGELOG.md
@@ -23,6 +23,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+*   Add `get_selected_dates_str` to TimeCoordinates.
+
 ## [2.0.0] - 2026-09-02
 
 *   **Breaking change**: Refactor `NotFittedModelError` to `common.errors.NotFittedModelError`.